Added a simple packaging script
The added script will generate a tar.bz2 file based on a git tag.
When using the 'tarbz2' target, it will make a package based on
the currently checked-out git tag.

If you try to run 'make tarbz2' on a commit which is not
identifiable as a tag, it will fail.  This is by design, to
ensure only tagged versions can result in a tar.bz2 file.

Signed-off-by: David Sommerseth <davids@redhat.com>