Add support for using previous results in meta-searching
Meta-searching on the server was working by searching all messages for
each search request and then applying a logical "and" on all the partial
results.  But this is optimized, and now subsequent search requests
include the sequence set of messages that have already been matched in
previous requests, and thus in theory improving performance and saving
resources on the server.