Add support for loading CA certificates from a file
Previously CA certificates could only be loaded from a directory.

If certificate verification fails a more strict policy is followed and
the connection fails instead of continuing to the local certificate file
(this reverts the previous commit 39ca3b8).  Also, the connection is now
dropped without asking the user, if there's a certificate mismatch.