Adding UTF-8 support conforming to RFC 6855
imapfilter handles UTF-8 encoding preprending CHARSET 'UTF-8' to queries when options.charset is set to 'UTF-8' in the configuration file.

This seems to work for Mailboxes that are transformed to UTF-7 according to RFC 3501, but not for other queries like the ones sent with send_query('X-GM-RAW ...').

For example:

  C (3): 1015 UID SEARCH CHARSET "UTF-8" ALL SUBJECT "Areté"
  S (3): 1015 BAD Could not parse command
  imapfilter: IMAP (3): 1015 BAD Could not parse command

  or

  C (3): 1016 UID SEARCH CHARSET "UTF-8" ALL X-GM-RAW "label:INBOX Areté"
  S (3): 1016 BAD Could not parse command
  imapfilter: IMAP (3): 1016 BAD Could not parse command

Since GMail accepts UTF-8 for X-GM-RAW queries, and UTF-8 in IMAP is supported since RFC 6855, this implementation of the UTF8=ACCEPT capability makes these queries work.
This also makes unnecessary the transcoding into UTF-7 for Mailboxes containing non-ASCII characters when the server accepts UTF-8 conforming to RFC 6855.