Parse the transfer encoding from the IMAP structure reply
Parses the transfer encoding field from the IMAP structure reply,
so that it can be accessed as transfer_encoding on the parts information field.

This field will contain the equivalent to the Content-Transfer-Encoding in the
raw multipart text.

This makes it possible for filters to automatically choose an appropritate
content decoder instead of relying on heuristics.